+++ Modmata.h
@@ -8,12 +8,12 @@ class ModmataClass
 {
   public:
     void begin();
-    void attach(uint16_t command, struct registers (*fn)(uint16_t argc, uint16_t *argv));
+    void attach(uint8_t command, struct registers (*fn)(uint8_t argc, uint8_t *argv));
     void processInput();
     bool available();
   
   private:
-    struct registers (*callbackFunctions[20])(uint16_t argc, uint16_t *argv);
+    struct registers (*callbackFunctions[20])(uint8_t argc, uint8_t *argv);
     ModbusSerial mb;
     
 };
